[REF] Re-organize Common lib
