+++ ScraperLib/IFetcher.cs
